api-service/requirements.txt
Flask==3.0.2
scikit-learn==1.2.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
